refactor(cowork): inline isQuestionLikeMemoryText and remove legacy memory modules
The coworkMemoryExtractor and coworkMemoryJudge modules were only
used by the removed yd_cowork engine. Inline the one still-needed
function (isQuestionLikeMemoryText) into coworkStore.ts and delete
the dead applyTurnMemoryUpdates method along with both modules.
